requirements.txt
astroid==2.3.3
autopep8==1.4.4
boto3==1.10.28
botocore==1.13.28
Django==2.2.7
djangorestframework==3.10.3
docutils==0.15.2
isort==4.3.21
jmespath==0.9.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.4
pycodestyle==2.5.0
pylint==2.4.4
python-dateutil==2.8.0
pytz==2019.3
s3transfer==0.2.1
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.7
wrapt==1.11.2
